feat: The type of result is changed from int to long

+++ src/main/java/cc/xpbootcamp/warmup/fibonacci/FibonacciCalculator.java
@@ -1,7 +1,7 @@
 package cc.xpbootcamp.warmup.fibonacci;
 
 public class FibonacciCalculator {
-    public int getFibonacciByPosition(int position) {
+    public long getFibonacciByPosition(int position) {
         if (position == 1 || position == 2)
             return 1;
